created BulkUploadReferences base class so that workplaces and staff references can extend from it
